+++ myUtils.cpp
@@ -5,22 +5,22 @@
 #include <math.h>
 #include <EEPROM.h>
 
-void saveValue(int index, double value) {
+void saveValue(int index, float value) {
   EEPROM.write(index, value);
   EEPROM.commit(); // EEPROM에 변경사항을 저장합니다.
 }
 
-double loadValue(int index) {
+float loadValue(int index) {
     return EEPROM.read(index);
 }
 
 // Moving average filter
-std::vector<double> movingAverageFilter(const std::vector<double>& data, int windowSize) {
-    std::vector<double> smoothedData(data.size());
+std::vector<float> movingAverageFilter(const std::vector<float>& data, int windowSize) {
+    std::vector<float> smoothedData(data.size());
     int halfWindow = windowSize / 2;
 
     for (size_t i = 0; i < data.size(); ++i) {
-        double sum = 0.0;
+        float sum = 0.0;
         int count = 0;
         for (int j = -halfWindow; j <= halfWindow; ++j) {
             if (i + j >= 0 && i + j < data.size()) {
@@ -35,7 +35,7 @@ std::vector<double> movingAverageFilter(const std::vector<double>& data, int win
 }
 
 // 극대 극소
-void findLocalExtrema(const std::vector<double>& data, std::vector<size_t>& maxima, std::vector<size_t>& minima) {
+void findLocalExtrema(const std::vector<float>& data, std::vector<size_t>& maxima, std::vector<size_t>& minima) {
     for (size_t i = 1; i < data.size() - 1; ++i) {
         if (data[i] > data[i - 1] && data[i] > data[i + 1]) {
             maxima.push_back(i);
@@ -46,9 +46,9 @@ void findLocalExtrema(const std::vector<double>& data, std::vector<size_t>& maxi
     }
 }
 
-std::vector<double> getLocalExtremeValue(const std::vector<double>& data, const int& windowSize) {
+std::vector<float> getLocalExtremeValue(const std::vector<float>& data, const int& windowSize) {
   // 이동 평균 필터 적용
-  std::vector<double> smoothedData = movingAverageFilter(data, windowSize);
+  std::vector<float> smoothedData = movingAverageFilter(data, windowSize);
 
   // 극댓값과 극솟값 저장 벡터
   std::vector<size_t> maxima;
@@ -57,8 +57,8 @@ std::vector<double> getLocalExtremeValue(const std::vector<double>& data, const 
   // 극값 찾기
   findLocalExtrema(smoothedData, maxima, minima);
 
-  double localMax = 0;
-  double localMin = 0;
+  float localMax = 0;
+  float localMin = 0;
   int numOfLocalMax = 0;
   int numOfLocalMin = 0;
 
@@ -83,7 +83,7 @@ std::vector<double> getLocalExtremeValue(const std::vector<double>& data, const 
   }
   Serial.println("============================================");
 
-  std::vector<double> result;
+  std::vector<float> result;
   result.push_back(localMax / numOfLocalMax);
   result.push_back(localMin / numOfLocalMin);
