+++ myUtils.hpp
@@ -3,10 +3,10 @@
 
 #include <vector>
 
-void saveValue(int index, double value);
-double loadValue(int index);
-std::vector<double> movingAverageFilter(const std::vector<double>& data, int windowSize);
-void findLocalExtrema(const std::vector<double>& data, std::vector<size_t>& maxima, std::vector<size_t>& minima);
-std::vector<double> getLocalExtremeValue(const std::vector<double>& data, const int& windowSize);
+void saveValue(int index, float value);
+float loadValue(int index);
+std::vector<float> movingAverageFilter(const std::vector<float>& data, int windowSize);
+void findLocalExtrema(const std::vector<float>& data, std::vector<size_t>& maxima, std::vector<size_t>& minima);
+std::vector<float> getLocalExtremeValue(const std::vector<float>& data, const int& windowSize);
 
 #endif
